requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.27
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
email-validator==2.1.0.post1
motor==3.4
pymongo==4.8.0
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.18.1
apscheduler==3.10.4
websockets==12.0
aiohttp==3.9.3
pytest==8.0.1
httpx==0.27.0
pandas==2.1.0
openpyxl==3.1.2 